// src/app/enum/key.enum.ts
// Copyright (c) 2026 Zwelithini Ngomane (cypriel17@gmail.com). All rights reserved.
// LKCentrix HR & Payroll Management System — ORION
// Unauthorised copying, distribution or modification is strictly prohibited.

export enum Key {
    TOKEN = '[KEY] TOKEN',
    REFRESH_TOKEN = '[REFRESH] REFRESH_TOKEN',
    THEME = '[THEME] PREFERENCE',
    COMPANY_ID = '[COMPANY] ID',
    SUPERADMIN_TOKEN = '[SUPERADMIN] TOKEN'
}